+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 2.0.0
+
+- Enhance support for imports and requires
+ - Only jump to files when an import is clicked instead of the variable it declared
+
 #### 1.0.0
 
 - Initial release